Document game design changes happening in the translation.

+++ sbbs3/scfgdefs.h
@@ -2,7 +2,7 @@
 
 /* Synchronet configuration structure (scfg_t) definition */
 
-/* $Id: scfgdefs.h,v 1.34 2009/11/11 05:30:06 rswindell Exp $ */
+/* $Id: scfgdefs.h,v 1.33 2009/03/20 09:52:58 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
@@ -118,7 +118,7 @@ typedef struct {							/* Transfer Library Information */
 				code_prefix[LEN_CODE+1],	/* Prefix for internal code */
 				parent_path[48];			/* Parent for dir paths */
 	uchar		*ar;
-	uint32_t	offline_dir;				/* Offline file directory */
+	uint16_t	offline_dir;				/* Offline file directory */
 
 } lib_t;
 
@@ -545,9 +545,9 @@ typedef struct
 	char			mods_dir[LEN_DIR+1];
 	char			logs_dir[LEN_DIR+1];
 	char			node_path[MAX_NODES][LEN_DIR+1]; /* paths to all node dirs */
-	uint			sysop_dir;			/* Destination for uploads to sysop */
-	uint			user_dir;			/* Directory for user to user xfers */
-	uint			upload_dir; 		/* Directory where all uploads go */
+	uint16_t		sysop_dir;			/* Destination for uploads to sysop */
+	uint16_t		user_dir;			/* Directory for user to user xfers */
+	uint16_t		upload_dir; 		/* Directory where all uploads go */
 	char **			altpath;			/* Alternate paths for files */
 	uint16_t		altpaths;			/* Total number of alternate paths */
 	uint16_t		leech_pct;			/* Leech detection percentage */
